Fix: KICK: exception handling - 방장이 스스로를 강퇴한 경우 방장이 다른 유저에게 상속되도록 변경 - 방장이 없는 유저를 강퇴하려 한 경우 에러 메세지가 전송되도록 변경

+++ refactoring/command.cpp
@@ -90,18 +90,32 @@ void kick(User &user, Channel &channel, vector<string> command, int fd)
     string clientNickName = user.getNickName(fd);
     string clientLoginName = user.getLoginName(fd);
     string clientHostName = user.getHostName(fd);
-    if (channel.isAdmin(channelName, fd))
+    string message;
+
+    if (channel.isAdmin(channelName, fd) && user.isExistUser(target))
     {
-        string message = prefixMessage(clientNickName, clientLoginName, clientHostName, "kick", channelName + " " + target + " :" + shortMessage);
+        message = prefixMessage(clientNickName, clientLoginName, clientHostName, "kick", channelName + " " + target + " " + shortMessage);
         channel.setBroadCastMessage(channelName, 0, message, user);
         channel.deleteUser(channelName, user.getUserFd(target));
         user.deleteChannel(user.getUserFd(target), channelName);
+        if (user.getUserFd(target) == fd)
+        {
+            vector<int> userList =  channel.getUserList_vec(channelName);
+            for (int i = 0; i < userList.size(); ++i)
+            {
+                int currentUser = userList[i];
+                string fullDiffMessage = serverMessage(353, user.getNickName(currentUser), user.getLoginName(currentUser), \
+                    channelName, channel.getUserList(user, channelName, currentUser));
+                user.setWriteBuffer(currentUser, fullDiffMessage);
+            }
+        }
+        return ;
     }
+    if (channel.isAdmin(channelName, fd))
+        message = serverMessage(401, clientNickName, clientLoginName, "" , "No such nick/channel");
     else
-    {
-        string message = serverMessage(481, clientNickName, clientLoginName, "" , "Permission Denied- You're not an IRC operator");
-        user.setWriteBuffer(fd, message);
-    }
+        message = serverMessage(481, clientNickName, clientLoginName, "" , "Permission Denied- You're not an IRC operator");
+    user.setWriteBuffer(fd, message);
 }
 
 void quit(User &user, Channel &channel, string message, int fd){
